Move actor exporting and loading code to function_manager.py
